update product manager

+++ models/transaction.models.js
@@ -11,15 +11,15 @@ module.exports = (sequelize, Sequelize) => {
             allowNull: false
         },
         total_discount: {
-            type: Sequelize.DOUBLE,
+            type: Sequelize.DECIMAL(18, 2),
             allowNull: true
         },
         total_amount: {
-            type: Sequelize.DOUBLE,
+            type: Sequelize.DECIMAL(18, 2),
             
         },
         total_payment:{
-            type: Sequelize.DOUBLE,
+            type: Sequelize.DECIMAL(18, 2),
             
         }
